fix Coverage.py reporting for expr macros, too
`_visit_expansion` used to `copy_location` to the top-level node from the
replaced target node. This approach works for expr macros, but not for block and
decorator macros.

For example for a block macro, the top-level code often already has correct
location info, pointing to source lines inside the `with` block. That's
obviously different from the line from which the macro was invoked. So if we
overwrite that location info, we run the risk of reporting some lines inside
the block as not covered (when we fill in missing location info automatically).

This commit moves the `copy_location` call to `core`, which is the layer that
defines the three kinds of macros. It's now called for expr macros only.

Block and decorator macros already use a different approach (from an earlier
commit), inserting a dummy statement for coverage reporting purposes, for each
macro invocation line that was compiled away. That has a small run-time cost (of
evaluating a constant None and discarding it), but the reported test coverage
should always be correct.

For any macro that successfully expanded, the invocation line is reported as
covered. For any lines inside the block, coverage will depend on whether the
code taken from that line actually ran, as it should.