improve macro expansion error reporting
We shouldn't discard the intermediate MacroExpansionErrors in nested macro
invocations, because the actual use site could be an expr macro invocation
inside a block macro invocation, and it may be important to see both (e.g.
if the block macro is an AST-walking code rewriter, that context is important
for understanding the failure).

This has the effect of somewhat uselessly reporting each level of expr macro
invocation when those are nested, but that's a minor inconvenience.

Also, if the unparsed code contains multiple lines, begin it on its own line.