h[]: support hygienically capturing macro names, too
The core keeps a global registry (across all modules being expanded) for those.
The quasiquote system fills it in when it detects `q[h[macroname][...]]`.

`macroname` must be in the current expander's bindings when the use site of `q`
(which is likely in another macro) is being expanded. Just like with run-time
values and `h[]`, it doesn't matter what macro (if any) that name is bound to at
the final use site (of that second macro that used `q`).