fix some bugs in source location reporting
We don't need `fill_root_location`, after all, just a `fix_missing_locations`
that can take the initial info from an AST node given as reference. The one in
the `ast` stdlib module always starts counting from line 1, column 0, which is
not applicable for the outermost layer of new nodes is a macro expansion.

The correct solution (whether macros expand outside in or inside out) is to fill
in missing location info only (relying on that macros don't populate it manually),
and use the macro invocation node as the reference.

This way, any new nodes added by a macro get their lineno set to the line that
has the macro invocation.

We still need to add an extra `Expr` node to always report correct coverage for
block and decorator macro invocations, because the macro invocation itself does
not necessarily correspond to any line in the output (e.g. if a macro only edits
existing nodes without adding any new ones).