Add error handling to `a`
Via the magic of a run-time check at the use site of `q`, now `a` accepts
only an expr node if invoked as `a[]` and only statement nodes if invoked
as `with a`.

This should cut down on the most common cause of mysterious compile errors
(i.e. getting bitten by Python's invisible `ast.Expr` node).

(How a run-time check can help that: it's run time for the macro definition that
uses `q` to build its output AST, while it's still macro-expansion time for its
use site. So as soon as one tries to invoke a macro that uses `a` incorrectly,
the check will trigger.)

The implementation is rather complex, but it's unlikely it could be simplified.