report filename in coverage dummy node
Useful in a chain of block macros from multiple files. Especially if `mymacro`
uses `with q`/`with a`, the human-readable info in the coverage dummy node of
`with a` will show the line number from the file in which `mymacro` is defined,
not in the file where it is used (even though in the actual AST, the line number
will be set correctly for the use site).

So now the human-readable info is more explicit.