standardize on double quotes for all strings (for now)
This fixes the mishmash of string quote styles, due to `mcpy` preferring single
quotes, and me preferring double quotes when writing new code.

We may revisit this later to use double quote for human-readable strings and
single quote for machine-readable strings, but that requires a full once-over
with some thought applied.