improve `fake_import`
Now it behaves like the real importer (actually using the `expand_ast` public
API function added in the previous commit).

Also, the source locations are now filled at the use site, by using the new
`mcpyrate.metatools.fill_location` macro.

The source location stuff could still be slightly improved, by adding a mode
to `q` to lift the original source locations into the astified tree (via the
`lineno` and `col_offset` kwargs of the astified node, like `fill_location`
now does). Most of the time we don't want to do that, but when the point is to
produce a run-time AST value to be compiled by `fake_import`, for that use case
it would be appropriate.