support nested quasiquotes properly, like Lisps do
Unquoting now only occurs when quote level hits zero. Inner quotes and unquotes
are detected, for tracking the quote level, but are then left in the output
as-is.

Note as-is means "as unexpanded macro invocations". Because the quasiquote
operators are just macros, and in macro-enabled Python, the tradition is that
a function actually being a macro is a property of the *use site*, not of its
definition site, it follows that there's no guarantee whether the quote
operators are in the expander's bindings at any later time. Even if they are,
there is no guarantee whether they still have the names they had at the time
when the outermost quote expanded.

However, this solution gives useful level separation that has real practical
applications; see the dynamically generated module example in `test_compiler.py`.

Not sure if this is a breaking change, because the previous behavior concerning
nested quotes didn't make any sense.

Idea for next-next-gen: maybe quasiquotes should be considered core
functionality (not just an add-on feature implemented with garden variety
macros, as it is now), and shouldn't support renaming at the use site (unlike
now, and unlike regular macros).